refactor(scripts): one definition of the cross-carrier rule comparison
Extract the three-way comparison (identical / provenance-only / body-diff)
and the rule-directory census into _lib/carrier_divergence.ts. Both were
about to gain a second copy: the advisory divergence report needs the
comparison, and the conformance scan needs the census.

Why a library rather than a second copy in each place:

- measure_scope_dedup asks how many installed twins a byte-identity dedup
  could skip; the report asks which rules a reader must look at. The
  directory walks are legitimately different and stay where they are. The
  comparison is not, and a second copy is how "differs only in the
  ownership stamp" and "differs in body" drift into meaning two things.
- censusRuleDir moved out of preamble_byte_census because that module pulls
  yaml plus the cold-start report, and importing that chain into a bundled
  CLI path to reach fifteen lines of statSync is the wrong trade. It is
  re-exported there, so every existing caller is unchanged.

No behaviour change: scope_dedup.test.ts (12 tests) passes untouched, and
the measure_scope_dedup output is byte-identical.

Executes part of road-to-conformance-round5 Phase 1.3.
